feat(SPV-848): fix lint errors
